Import notes from pagerduty incidents
Pagerduty allows notes to be recorded on incidents. This is often
more convenient for the on-call than bringing up opsweeksly and
generating the report to add a note.
With this change opsweekly entries that are fetched from pagerduty
will include the notes from the pagerduty incident. Once an
entry is saved in opsweekly, further updates to the pagerduty notes
will not be imported.

The pagerduty api only allows fetching the notes for a single
incident at a time, so for large reports this feature can
considerably slow down the report generation. However we deem
this performance hit to be outweighed by the productivity
benefits of allowing the on call to record their notes on
the incidents in PD as they handle the incidents.